@@ -1,13 +1,13 @@
-import React, { useState, useMemo } from 'react';
-import { Link, useParams } from 'react-router-dom';
+import { useState, useMemo } from 'react';
+import { Link } from 'react-router-dom';
 import Card, { CardInfo } from '../Card/Card';
 import styles from './employeecards.module.css';
 import { clock, phone, wheel, user } from '../../icons/userInfo/index';
-import { Employee, Admin, Driver } from '../../types';
+import { Employee } from '../../types';
 import { useEmployees } from '../../context/EmployeesContext';
-import formatAvailability from '../../util/employee';
 import { AdminType } from '../../../../server/src/models/admin';
 import { DriverType } from '../../../../server/src/models/driver';
+import { Button } from '../FormElements/FormElements';
 
 const formatPhone = (phoneNumber: string) => {
   const areaCode = phoneNumber.substring(0, 3);
@@ -20,45 +20,34 @@ type EmployeeCardProps = {
   id: string;
   employee: Employee;
 };
-type EmployeeDetailProps = {
-  id: string;
-  firstName: string;
-  lastName: string;
-  role?: string[];
-  driverId?: string;
-  netId: string;
-  phone: string;
-  availability?: string[][];
-  photoLink?: string;
-  startDate?: string;
-};
-//Convert DriverType to EmployeeType
-const DriverToEmployees = (drivers: DriverType[]): EmployeeDetailProps[] => {
-  return drivers.map((driver) => ({
-    id: driver.id,
-    firstName: driver.firstName,
-    lastName: driver.lastName,
-    availability: formatAvailability(driver.availability)!,
-    netId: driver.email.split('@')[0],
-    phone: driver.phoneNumber,
-    photoLink: driver.photoLink,
-    startDate: driver.startDate,
-  }));
-};
 
-//Convert AdminType to EmployeeType
-const AdminToEmployees = (admins: AdminType[]): EmployeeDetailProps[] => {
-  return admins.map((admin) => ({
-    id: admin.id,
-    firstName: admin.firstName,
-    lastName: admin.lastName,
-    type: admin.type,
-    isDriver: admin.isDriver,
-    netId: admin.email.split('@')[0],
-    phone: admin.phoneNumber,
-    photoLink: admin.photoLink,
-  }));
-};
+//Convert DriverType to EmployeeType
+// const DriverToEmployees = (drivers: DriverType[]): EmployeeDetailProps[] => {
+//   return drivers.map((driver) => ({
+//     id: driver.id,
+//     firstName: driver.firstName,
+//     lastName: driver.lastName,
+//     availability: formatAvailability(driver.availability)!,
+//     netId: driver.email.split('@')[0],
+//     phone: driver.phoneNumber,
+//     photoLink: driver.photoLink,
+//     startDate: driver.startDate,
+//   }));
+// };
+
+// //Convert AdminType to EmployeeType
+// const AdminToEmployees = (admins: AdminType[]): EmployeeDetailProps[] => {
+//   return admins.map((admin) => ({
+//     id: admin.id,
+//     firstName: admin.firstName,
+//     lastName: admin.lastName,
+//     type: admin.type,
+//     isDriver: admin.isDriver,
+//     netId: admin.email.split('@')[0],
+//     phone: admin.phoneNumber,
+//     photoLink: admin.photoLink,
+//   }));
+// };
 
 const EmployeeCard = ({
   id,
@@ -85,7 +74,7 @@ const EmployeeCard = ({
     }
 
     return Object.entries(availability)
-      .filter(([day, timeRange]) => timeRange?.startTime && timeRange?.endTime)
+      .filter(([_, timeRange]) => timeRange?.startTime && timeRange?.endTime)
       .map(
         ([day, timeRange]) =>
           `${day}: ${timeRange.startTime} - ${timeRange.endTime}`
@@ -160,10 +149,10 @@ const searchableFields = (employee: DriverType | AdminType) => {
     employee.email,
     employee.phoneNumber,
   ];
-  if ('vehicle' in employee) {
+  if ('vehicle' in employee && employee.vehicle) {
     fields.push(employee.vehicle.name);
   }
-  if ('type' in employee) {
+  if ('type' in employee && employee.type) {
     fields.push(...employee.type);
   }
   return fields;
@@ -183,6 +172,8 @@ type EmployeeCardsProps = {
 
 const EmployeeCards = ({ query }: EmployeeCardsProps) => {
   const { admins, drivers } = useEmployees();
+  const [filterAdmin, setFilterAdmin] = useState(false);
+  const [filterDriver, setFilterDriver] = useState(false);
 
   const employees = useMemo(() => {
     const allEmployees = [...admins, ...drivers];
@@ -190,22 +181,55 @@ const EmployeeCards = ({ query }: EmployeeCardsProps) => {
     allEmployees.forEach((employee) => {
       employeeSet[employee.id] = { ...employeeSet[employee.id], ...employee };
     });
-    const sortedEmployees = Object.values(employeeSet).sort(
-      (a: Employee, b: Employee) => a.firstName.localeCompare(b.firstName)
-    );
+    const sortedEmployees = Object.values(employeeSet)
+      .filter((employee) => {
+        // if both or neither filters are selected, show all employees
+        if (filterAdmin == filterDriver) {
+          return true;
+        }
+        if (filterDriver) {
+          return 'availability' in employee;
+        }
+        if (filterAdmin) {
+          return 'type' in employee;
+        }
+      })
+      .sort((a, b) => a.firstName.localeCompare(b.firstName));
     if (!query) {
       return sortedEmployees;
     }
     // By filtering after coalescing step, we keep role info intact
     return sortedEmployees.filter(matchesQuery(query));
-  }, [admins, drivers, query]);
+  }, [admins, drivers, query, filterAdmin, filterDriver]);
 
   return (
-    <div className={styles.cardsContainer}>
-      {employees.map((employee) => (
-        <EmployeeCard key={employee.id} id={employee.id} employee={employee} />
-      ))}
-    </div>
+    <>
+      <div className={styles.filtersContainer}>
+        <Button
+          type="button"
+          onClick={() => setFilterDriver((v) => !v)}
+          outline={!filterDriver}
+        >
+          Drivers
+        </Button>
+        <Button
+          type="button"
+          onClick={() => setFilterAdmin((v) => !v)}
+          outline={!filterAdmin}
+        >
+          Admins
+        </Button>
+      </div>
+      <div className={styles.cardsContainer}>
+        {employees.map((employee) => (
+          <EmployeeCard
+            key={employee.id}
+            id={employee.id}
+            employee={employee}
+          />
+        ))}
+      </div>
+    </>
   );
 };
 
