@@ -22,55 +22,50 @@ const EmployeeInfo = ({
   return (
     <div className={styles.inputContainer}>
       <div className={styles.col1}>
-        <Label htmlFor={'firstName'}>First Name</Label>
+        <Label htmlFor="firstName">First Name</Label>
         <Input
           id="firstName"
-          name="firstName"
           type="text"
           defaultValue={firstName}
           className={cn(styles.input)}
           aria-required="true"
-          ref={register({ required: true })}
+          {...register('firstName', { required: true })}
         />
         {errors.firstName && (
           <p className={styles.error}>Please enter a valid name</p>
         )}
       </div>
       <div className={styles.col2}>
-        <Label htmlFor={'lastName'}>Last Name</Label>
+        <Label htmlFor="lastName">Last Name</Label>
         <Input
           id="lastName"
-          name="lastName"
           type="text"
           defaultValue={lastName}
           className={cn(styles.input)}
           aria-required="true"
-          ref={register({ required: true })}
+          {...register('lastName', { required: true })}
         />
         {errors.lastName && (
           <p className={styles.error}>Please enter a valid name</p>
         )}
       </div>
       <div className={styles.col1}>
-        <Label htmlFor={'netid'}>NetID</Label>
+        <Label htmlFor="netid">NetID</Label>
         <Input
-          name="netid"
           id="netid"
           type="text"
           defaultValue={netId}
           className={cn(styles.input)}
           aria-required="true"
-          ref={register({ required: true })}
+          {...register('netid', { required: true })}
         />
         {errors.netid && (
           <p className={styles.error}>Please enter a valid NetID</p>
         )}
       </div>
-
       <div className={styles.col2}>
-        <Label htmlFor={'phoneNumber'}>Phone Number</Label>
+        <Label htmlFor="phoneNumber">Phone Number</Label>
         <Input
-          name="phoneNumber"
           id="phoneNumber"
           type="tel"
           defaultValue={phone}
@@ -78,15 +73,15 @@ const EmployeeInfo = ({
           max={10}
           aria-required="true"
           className={cn(styles.input)}
-          ref={register({
+          {...register('phoneNumber', {
             required: true,
-            pattern: /[0-9]{10}/,
+            pattern: /^[0-9]{10}$/,
             maxLength: 10,
             minLength: 10,
           })}
         />
-        {formState.errors.phoneNumber && (
-          <p className={styles.error}>Please enter a valid phone number </p>
+        {errors.phoneNumber && (
+          <p className={styles.error}>Please enter a valid phone number</p>
         )}
       </div>
     </div>
