@@ -35,6 +35,8 @@ type ModalProps = {
   onClose?: () => void;
   displayClose?: boolean;
   isRider?: boolean;
+  id?: string;
+  arialabelledby?: string;
 };
 
 const Modal = ({
@@ -46,6 +48,8 @@ const Modal = ({
   onClose,
   displayClose,
   isRider = true,
+  arialabelledby,
+  id = 'modal',
 }: ModalProps) => {
   // Wrapping children in Array to match type for numPages
   const pages = paginate ? (children as React.ReactNodeArray) : [children];
@@ -69,12 +73,21 @@ const Modal = ({
             }}
           >
             <div className={styles.background}>
-              <div className={styles.modal}>
+              <div
+                className={styles.modal}
+                role="dialog"
+                aria-modal="true"
+                aria-labelledby={arialabelledby ?? id}
+              >
                 <div className={styles.topContainer}>
                   {isRider ? (
-                    <h1 className={styles.title}>{currentTitle}</h1>
+                    <h1 className={styles.title} id={id}>
+                      {currentTitle}
+                    </h1>
                   ) : (
-                    <div className={styles.title}>{currentTitle}</div>
+                    <div className={styles.title} id={id}>
+                      {currentTitle}
+                    </div>
                   )}
                   {!displayClose && isRider && (
                     <button
