@@ -5,13 +5,25 @@ import styles from '../ridemodal.module.css';
 import { Label, Input, Button } from '../../FormElements/FormElements';
 import axios from '../../../util/axios';
 
-const DriverPage = ({ onBack, onSubmit, formData }: ModalPageProps) => {
-  const { register, handleSubmit, formState } = useForm({
+interface FormData {
+  driver: string;
+}
+
+const DriverPage = ({
+  onBack,
+  onSubmit,
+  formData,
+  labelid,
+}: ModalPageProps & { labelid?: string }) => {
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<FormData>({
     defaultValues: {
       driver: formData?.driver ?? '',
     },
   });
-  const { errors } = formState;
   const [loaded, setLoaded] = useState(false);
 
   const { date, pickupTime: startTime, dropoffTime: endTime } = formData!;
@@ -38,7 +50,12 @@ const DriverPage = ({ onBack, onSubmit, formData }: ModalPageProps) => {
   return (
     <form onSubmit={handleSubmit(onSubmit)} className={styles.form}>
       <div className={styles.inputContainer}>
-        <div className={styles.drivers}>
+        <div
+          className={styles.drivers}
+          aria-required="true"
+          role="radiogroup"
+          aria-labelledby={labelid}
+        >
           {loaded ? (
             availableDrivers.map((d) => (
               <div className={styles.driver} key={d.id}>
@@ -51,11 +68,9 @@ const DriverPage = ({ onBack, onSubmit, formData }: ModalPageProps) => {
                 <Input
                   id={d.firstName + d.lastName}
                   className={styles.driverRadio}
-                  name="driver"
                   type="radio"
                   value={d.id}
-                  ref={register({ required: true })}
-                  aria-required="true"
+                  {...register('driver', { required: true })}
                 />
               </div>
             ))
