@@ -100,11 +100,24 @@ const RideModal = ({ open, close, ride, editSingle }: RideModalProps) => {
 
   const submitData = () => setIsSubmitted(true);
 
+  /**
+   * Converts a ride that repeats into a number array representation used by
+   * the internal representation of a ride
+   *
+   * @param date a string representation of the ride start date
+   * @param repeats an enum representing how often this ride repeats: Daily,
+   * Weekly, or Custom
+   * @param days Used if the ride repeats on custom days. An object that
+   * maps days (Mon, Tue, Wed, Thur, Fri) to strings, where the string value is
+   * non-empty if the ride repeats on that day
+   * @returns a number array containing the days of the week where the ride repeats,
+   * with Monday represented as 1, Tuesday represented as 2, etc.
+   */
   const getRecurringDays = (
     date: string,
     repeats: RepeatValues,
     days: ObjectType
-  ) => {
+  ): number[] => {
     switch (repeats) {
       case RepeatValues.Daily:
         return [1, 2, 3, 4, 5];
@@ -227,6 +240,7 @@ const RideModal = ({ open, close, ride, editSingle }: RideModalProps) => {
         isOpen={isOpen}
         currentPage={currentPage}
         onClose={closeModal}
+        id="ride-modal"
       >
         <RideTimesPage
           formData={formData}
@@ -236,6 +250,7 @@ const RideModal = ({ open, close, ride, editSingle }: RideModalProps) => {
           formData={formData}
           onBack={goPrevPage}
           onSubmit={saveDataThen(goNextPage)}
+          labelid="ride-modal"
         />
         <RiderInfoPage
           formData={formData}
