@@ -1,6 +1,17 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { Calendar, momentLocalizer } from 'react-big-calendar';
-import withDragAndDrop from 'react-big-calendar/lib/addons/dragAndDrop';
+import React, {
+  ComponentType,
+  JSXElementConstructor,
+  PropsWithChildren,
+  ReactElement,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react';
+import {
+  Calendar,
+  EventWrapperProps,
+  momentLocalizer,
+} from 'react-big-calendar';
 import cn from 'classnames';
 import moment from 'moment';
 import { Ride, Driver } from '../../types';
@@ -12,6 +23,7 @@ import Modal from '../RideStatus/SModal';
 import { useEmployees } from '../../context/EmployeesContext';
 import { useRides } from '../../context/RidesContext';
 import axios from '../../util/axios';
+import { newDate } from 'react-datepicker/dist/date_utils';
 
 const colorMap = new Map<string, string[]>([
   ['red', ['FFA26B', 'FFC7A6']],
@@ -152,6 +164,36 @@ Rider: ${ride.rider.firstName} ${ride.rider.lastName}`,
     setCurrentRide(event.ride);
   };
 
+  const TabbableEventWrapper: ComponentType<
+    PropsWithChildren<EventWrapperProps<CalEvent>>
+  > = useMemo(
+    () => (props) => {
+      const child = React.Children.only(props.children) as ReactElement<
+        any,
+        string | JSXElementConstructor<any>
+      >;
+      return (
+        <div>
+          {React.cloneElement(child, {
+            tabIndex: 0,
+            onKeyDown: (event: React.KeyboardEvent<HTMLDivElement>): void => {
+              if (
+                event.key === 'Enter' ||
+                event.key === ' ' ||
+                event.key === 'Spacebar'
+              ) {
+                // Prevents spacebar from scroll event
+                event.preventDefault();
+                onSelectEvent(props.event);
+              }
+            },
+          })}
+        </div>
+      );
+    },
+    []
+  );
+
   const handleChangeViewState = () => setViewMore(!viewMore);
 
   return (
@@ -173,6 +215,7 @@ Rider: ${ride.rider.firstName} ${ride.rider.lastName}`,
             localizer={localizer}
             toolbar={false}
             step={5}
+            defaultDate={scheduleDay}
             timeslots={12}
             showMultiDayTimes={true}
             events={events}
@@ -180,12 +223,12 @@ Rider: ${ride.rider.firstName} ${ride.rider.lastName}`,
             onSelectEvent={onSelectEvent}
             min={minTime}
             max={maxTime}
-            date={scheduleDay}
             resources={calDrivers}
             resourceIdAccessor="resourceId"
             resourceTitleAccessor="resourceTitle"
             eventPropGetter={eventStyle}
             slotPropGetter={slotStyle}
+            components={{ eventWrapper: TabbableEventWrapper }}
           />
         </div>
       </div>
